Split DatabaseService into EntryStore, Classification, Upgrade, Import, Recovery, and FileOperations
